airflow/requirements.txt
apache-airflow==2.8.0
pyspark==3.5.0
delta-spark==3.0.0
requests==2.31.0
pip==22.0.2
apache-airflow-providers-apache-spark==4.6.0
psycopg2==2.9.9


